backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==45.0.3
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.0.0
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-slugify==8.0.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
typing_extensions==4.12.2
uvicorn==0.32.1
